数据分片

+++ sharding-jdbc-range-group/src/main/java/com/shardingjdbc/service/UserService.java
@@ -1,15 +1,12 @@
 package com.shardingjdbc.service;
 
-import com.shardingjdbc.entity.User;
 
+import com.shardingjdbc.entity.User;
 import java.util.List;
 
-/**
- * @author yangbin
- */
 public interface UserService {
 
-    Long addUser(User user);
+    Integer addUser(User user);
 
     List<User> list();
 }
